refactor(core): add apis for http module
