+++ fhe_http_core/src/lib.rs
@@ -6,6 +6,7 @@ pub mod configs {
 pub mod utils {
     pub mod base64;
     pub mod file_ctl;
+    pub mod http;
     pub mod json;
 }
