+++ fhe_http_core/src/utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod base64;
 pub mod file_ctl;
+pub mod http;
 pub mod json;
